refactor(embed): extract chunk-blob store + sync-meta coordination into embed/chunk_store.rs
